requirements.txt
icecream
moviepy
openai
python-magic
rich
spacy
tiktoken
typer
